Default override virtlogd / virtlockd
Restarting virtlogd / virtlockd is really bad as guests would be fenced
and restarted. This is intentionally skipped on the package upgrade [1]
and tried to be the last that dies by upstream [2][3].

We should at least not preselect those services to avoid these issues,
therefore add them to the default override config.

[1]: https://salsa.debian.org/libvirt-team/libvirt/-/blob/debian/master/debian/rules#L254
[2]: https://libvirt.org/git/?p=libvirt.git;a=blob;f=src/logging/virtlogd.service.in;h=8ab547851772298ff570487f1ce2d3680c7e4d68;hb=HEAD#l13
[3]: https://libvirt.org/git/?p=libvirt.git;a=blob;f=src/locking/virtlockd.service.in;h=4a6fab05ab309a04a8565a5c73cbcfb47dae8d5d;hb=HEAD#l13

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>