Detect and report outdated VMs
Needrestart is awesome in detecting long running services or user
sessions that are still running with old binaries. But there is
another common kind of long-running process - VM Guests.

Detecting the common qemu/KVM guests and reporting those to the
user will help them to realize that also those guests should be
considered to be restarted (or migrated onto an updated host)

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>