properly detect dhclient
We make the call here that dhclient should be *detected*, if not
properly restarted. We try to workaround the problem that restarts
fail or destroy the network by adding ifupdown to the long list of
services that cannot be restarted.

We've been running an equivalent configuration in production at the
Tor Project over about 100 Debian machines running Debian buster and
bullseye for months without negative issues.

Closes: #225