mark unavailable firmware as CURRENT
This changes the policy of reporting missing firmware updates as
"UNKNOWN". Now, if there's no available firmware, we report
"current". That fixes needrestart on platforms that do not have
firmware support (#220) or platforms that *are* supported but for
which a CPU is to new to have an updated firmware available (#149).

It does *not* create a new "unavailable" status (#274) as this maps
poorly in Nagios reporting, but that is an option that was considered
here as well.

Closes: #149
Closes: #274
Closes: #220