OpenMetrics: metric types not accepted by prometheus (Closes: #310)
Apparently, prometheus does not implement the full OpenMetrics spec and
only supports a very limited set of metric types.

If the output in OpenMetrics format is going to be useful to people, we
need the output to be ingestible by prometheus since it's one of the
most popular openmetrics-type TSDB software.

* type info is trivially converted into a gauge
* stateset needs to be reworked into a gauge that has a label that
  changes to indicate the status

This patch also moves the logic about how status codes are interpreted
up closer to where they are being used. It keeps the output section
easier to read.